Organize PNG app/mime icons by size & role
Reorganize the way the application and MIME icons are stored:
- create subdirectories for each size
- create per-size subdirectories according to the role ("apps" or
  "mimetypes")
- move the icons by size in the proper subdirectory
- rename the application icons as "leocad"
- rename the MIME icons as "application-vnd.leocad"

This is done in preparation for installing all these icons: since qmake
does not support renaming files when installing them, then place them
in subdirectories so they are in the right directory structure of the
hicolor icon theme.

The name of the MIME icons follows the name of the MIME type they refer
to.